refactor: improve map settings layout
