--- libs/safe/src/lib/components/widgets/map-settings/map-properties/map-properties.component.scss
@@ -1,10 +1,0 @@
-safe-map, .map-container {
-  height: 450px;
-  width: 100%;
-}
-
-@media (max-width: 779px) {
-  .map-container{
-    height: 300px !important;
-  }
-}
