+++ libs/safe/src/lib/components/widgets/map-settings/map-properties/map-properties.component.ts
@@ -1,12 +1,7 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, Input } from '@angular/core';
 import { UntypedFormGroup } from '@angular/forms';
 import { SafeUnsubscribeComponent } from '../../../utils/unsubscribe/unsubscribe.component';
-import { takeUntil } from 'rxjs/operators';
-import {
-  MapConstructorSettings,
-  MapEvent,
-  MapEventType,
-} from '../../../ui/map/interfaces/map.interface';
+import { MapConstructorSettings } from '../../../ui/map/interfaces/map.interface';
 import { BASEMAPS } from '../../../ui/map/const/baseMaps';
 
 /**
@@ -17,13 +12,10 @@ import { BASEMAPS } from '../../../ui/map/const/baseMaps';
   templateUrl: './map-properties.component.html',
   styleUrls: ['./map-properties.component.scss'],
 })
-export class MapPropertiesComponent
-  extends SafeUnsubscribeComponent
-  implements OnInit
-{
+export class MapPropertiesComponent extends SafeUnsubscribeComponent {
   @Input() form!: UntypedFormGroup;
+  @Input() mapSettings!: MapConstructorSettings;
 
-  public mapSettings!: MapConstructorSettings;
   public baseMaps = BASEMAPS;
 
   /**
@@ -31,75 +23,6 @@ export class MapPropertiesComponent
    */
   constructor() {
     super();
-  }
-
-  /**
-   * Subscribe to settings changes to update map.
-   */
-  ngOnInit(): void {
-    const defaultMapSettings = {
-      basemap: this.form.value.basemap,
-      zoom: this.form.value.zoom,
-      centerLat: this.form.value.centerLat,
-      centerLong: this.form.value.centerLong,
-      timeDimension: this.form.value.timeDimension,
-    };
-    this.updateMapSettings(defaultMapSettings);
-    this.setUpFormListeners();
-  }
-
-  /**
-   * Set form listeners
-   */
-  private setUpFormListeners() {
-    this.form
-      .get('zoom')
-      ?.valueChanges.pipe(takeUntil(this.destroy$))
-      .subscribe((value) =>
-        this.updateMapSettings({ zoom: value } as MapConstructorSettings)
-      );
-    this.form
-      .get('centerLat')
-      ?.valueChanges.pipe(takeUntil(this.destroy$))
-      .subscribe((value) =>
-        this.updateMapSettings({ centerLat: value } as MapConstructorSettings)
-      );
-    this.form
-      .get('centerLong')
-      ?.valueChanges.pipe(takeUntil(this.destroy$))
-      .subscribe((value) =>
-        this.updateMapSettings({ centerLong: value } as MapConstructorSettings)
-      );
-    this.form
-      .get('basemap')
-      ?.valueChanges.pipe(takeUntil(this.destroy$))
-      .subscribe((value) =>
-        this.updateMapSettings({ basemap: value } as MapConstructorSettings)
-      );
-    this.form
-      .get('timeDimension')
-      ?.valueChanges.pipe(takeUntil(this.destroy$))
-      .subscribe((value) => {
-        this.updateMapSettings({
-          timeDimension: value,
-        } as MapConstructorSettings);
-      });
-  }
-
-  /**
-   * Update map settings
-   *
-   * @param settings new settings
-   */
-  private updateMapSettings(settings: MapConstructorSettings) {
-    if (this.mapSettings) {
-      this.mapSettings = {
-        ...this.mapSettings,
-        ...settings,
-      };
-    } else {
-      this.mapSettings = settings;
-    }
   }
 
   /**
@@ -115,27 +38,5 @@ export class MapPropertiesComponent
     this.form
       .get('zoom')
       ?.setValue(this.mapSettings.zoom, { emitEvent: false });
-  }
-
-  /**
-   * Handle leaflet map events
-   *
-   * @param event leaflet map event
-   */
-  handleMapEvent(event: MapEvent) {
-    switch (event.type) {
-      case MapEventType.MOVE_END:
-        this.mapSettings.centerLat = event.content.center.lat;
-        this.mapSettings.centerLong = event.content.center.lng;
-        break;
-      case MapEventType.ZOOM_END:
-        this.mapSettings.zoom = event.content.zoom;
-        this.form
-          .get('zoom')
-          ?.setValue(this.mapSettings.zoom, { emitEvent: false });
-        break;
-      default:
-        break;
-    }
   }
 }
